@@ -1,15 +1,12 @@
 <Project Sdk="Microsoft.NET.Sdk">
-
 	<PropertyGroup>
 		<BaseOutputPath>../../bin</BaseOutputPath>
 		<OutputType>library</OutputType>
-		<TargetFramework>net6.0</TargetFramework>
+		<TargetFramework>net8.0</TargetFramework>
 		<Nullable>enable</Nullable>
 		<GenerateAssemblyInfo>false</GenerateAssemblyInfo>
 	</PropertyGroup>
-
 	<ItemGroup>
-	  <ProjectReference Include="..\TensorKit\TensorKit.csproj" />
+		<ProjectReference Include="..\TensorKit\TensorKit.csproj" />
 	</ItemGroup>
-
 </Project>
