@@ -1,17 +1,15 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-
 	<PropertyGroup>
 		<BaseOutputPath>../../bin</BaseOutputPath>
 		<OutputType>library</OutputType>
-		<TargetFramework>net6.0</TargetFramework>
+		<TargetFramework>net8.0</TargetFramework>
 		<Nullable>enable</Nullable>
 		<GenerateAssemblyInfo>false</GenerateAssemblyInfo>
 	</PropertyGroup>
-
 	<ItemGroup>
-	  <Compile Remove="Model\ContextInfoData\**" />
-	  <EmbeddedResource Remove="Model\ContextInfoData\**" />
-	  <None Remove="Model\ContextInfoData\**" />
+		<Compile Remove="Model\ContextInfoData\**" />
+		<EmbeddedResource Remove="Model\ContextInfoData\**" />
+		<None Remove="Model\ContextInfoData\**" />
 	</ItemGroup>
 	<ItemGroup>
 		<ProjectReference Include="../LoggerKit/LoggerKit.csproj" />
@@ -20,7 +18,6 @@
 		<ProjectReference Include="..\TensorKit\TensorKit.csproj" />
 	</ItemGroup>
 	<ItemGroup>
-	  <Folder Include="ContextData/Naming/" />
+		<Folder Include="ContextData/Naming/" />
 	</ItemGroup>
-
 </Project>
