@@ -16,7 +16,7 @@ public class ContextInfoFillerEmptyData<TTensor> : IContextInfoFiller<TTensor>
     private readonly ITensorBuilder _keyBuilder;
     private readonly IContextClassifier<ContextInfo> _wordRoleClassifier;
     private readonly IEmptyDimensionClassifier _emptyDimensionClassifier;
-    private readonly IFakeDimensionClassifier _FakeDimensionClassifier;
+    private readonly IFakeDimensionClassifier _fakeDimensionClassifier;
 
     public ContextInfoFillerEmptyData(ITensorFactory<TTensor> keyFactory, ITensorBuilder keyBuilder, IAppOptionsStore appOptionsStore)
     {
@@ -24,31 +24,31 @@ public class ContextInfoFillerEmptyData<TTensor> : IContextInfoFiller<TTensor>
         _keyBuilder = keyBuilder;
         _wordRoleClassifier = appOptionsStore.GetOptions<IContextClassifier<ContextInfo>>();
         _emptyDimensionClassifier = appOptionsStore.GetOptions<IEmptyDimensionClassifier>();
-        _FakeDimensionClassifier = appOptionsStore.GetOptions<IFakeDimensionClassifier>();
+        _fakeDimensionClassifier = appOptionsStore.GetOptions<IFakeDimensionClassifier>();
     }
 
     public int Order { get; } = int.MaxValue;
 
     // context: ContextInfo, ContextInfoMatrix, build
-    public void Fill(IContextInfoDataset<ContextInfo, TTensor> contextInfoData, List<ContextInfo> elements, ExportMatrixOptions _matrixOptions)
+    public void Fill(IContextInfoDataset<ContextInfo, TTensor> contextInfoData, List<ContextInfo> elements, ExportMatrixOptions matrixOptions)
     {
-        if (!_matrixOptions.IncludeAllStandardActions)
+        if (!matrixOptions.IncludeAllStandardActions)
             return;
 
-        var allVerbs = _wordRoleClassifier.GetCombinedVerbs(elements.SelectMany(e => e.Contexts).Where(c => _wordRoleClassifier.IsVerb(c, _FakeDimensionClassifier)).Distinct().ToList()).ToList();
-        var allNouns = elements.SelectMany(e => e.Contexts).Where(c => _wordRoleClassifier.IsNoun(c, _FakeDimensionClassifier)).Distinct().ToList();
+        var allVerbs = _wordRoleClassifier.GetCombinedVerbs(elements.SelectMany(e => e.Contexts).Where(c => _wordRoleClassifier.IsVerb(c, _fakeDimensionClassifier)).Distinct()).ToList();
+        var allNouns = elements.SelectMany(e => e.Contexts).Where(c => _wordRoleClassifier.IsNoun(c, _fakeDimensionClassifier)).Distinct().ToList();
 
-        bool includeUnclassified = _matrixOptions.UnclassifiedPriority != UnclassifiedPriorityType.None;
+        bool includeUnclassified = matrixOptions.UnclassifiedPriority != UnclassifiedPriorityType.None;
 
         var nounsToUse = includeUnclassified
-            ? allNouns.Append(_emptyDimensionClassifier.EmptyDomain)
-            : allNouns;
+            ? allNouns.Append(_emptyDimensionClassifier.EmptyDomain).ToList()
+            : allNouns.ToList();
 
         foreach (var row in allVerbs)
         {
             foreach (var col in nounsToUse)
             {
-                var key = _keyBuilder.BuildTensor(TensorPermutationType.Standard, new[] { row, col }, _keyFactory.Create);
+                var key = _keyBuilder.BuildTensor(TensorPermutationType.Standard, [row, col], _keyFactory.Create);
                 contextInfoData.Add(null, key);
             }
         }
@@ -57,11 +57,11 @@ public class ContextInfoFillerEmptyData<TTensor> : IContextInfoFiller<TTensor>
         {
             foreach (var noun in allNouns)
             {
-                var key = _keyBuilder.BuildTensor(TensorPermutationType.Standard, new[] { _emptyDimensionClassifier.EmptyAction, noun }, _keyFactory.Create);
+                var key = _keyBuilder.BuildTensor(TensorPermutationType.Standard, [_emptyDimensionClassifier.EmptyAction, noun], _keyFactory.Create);
                 contextInfoData.Add(null, key);
             }
 
-            var emptyKey = _keyBuilder.BuildTensor(TensorPermutationType.Standard, new[] { _emptyDimensionClassifier.EmptyAction, _emptyDimensionClassifier.EmptyDomain }, _keyFactory.Create);
+            var emptyKey = _keyBuilder.BuildTensor(TensorPermutationType.Standard, [_emptyDimensionClassifier.EmptyAction, _emptyDimensionClassifier.EmptyDomain], _keyFactory.Create);
             contextInfoData.Add(null, emptyKey);
         }
     }
