@@ -35,7 +35,7 @@ public class HtmlMatrixSummaryBuilder<TTensor> : IHtmlMatrixSummaryBuilder<TTens
     {
         return uiMatrix.cols.DistinctBy(c => c.LabeledData).ToDictionary(col => col.LabeledData, domain => uiMatrix.rows.Sum(action =>
         {
-            var key = _keyBuilder.BuildTensor(orientation, new[] { action.LabeledData, domain.LabeledData }, _keyFactory.Create);
+            var key = _keyBuilder.BuildTensor(orientation, [action.LabeledData, domain.LabeledData], _keyFactory.Create);
             return matrix.TryGetValue(key, out var methods) ? methods.Count : 0;
         }));
     }
@@ -44,7 +44,7 @@ public class HtmlMatrixSummaryBuilder<TTensor> : IHtmlMatrixSummaryBuilder<TTens
     {
         return uiMatrix.rows.DistinctBy(r => r.LabeledData).ToDictionary(row => row.LabeledData, action => uiMatrix.cols.Sum(domain =>
         {
-            var key = _keyBuilder.BuildTensor(orientation, new[] { action.LabeledData, domain.LabeledData }, _keyFactory.Create);
+            var key = _keyBuilder.BuildTensor(orientation, [action.LabeledData, domain.LabeledData], _keyFactory.Create);
             return matrix.TryGetValue(key, out var methods) ? methods.Count : 0;
         }));
     }
