@@ -1,7 +1,6 @@
 ﻿namespace TensorKit.Model;
 
-public interface ITensorClassifier<TDimensionType, TData>
-    where TDimensionType : ITensorDimensionType
+public interface ITensorClassifier< TData>
     where TData : notnull
 {
     bool IsDimensionApplicable(TData ctx, string? dimensionName, int dimensionType);
